requirements-3.6.txt
apipkg==1.4
appnope==0.1.0
APScheduler==3.5.1
argh==0.26.2
asyncio-redis==0.14.3
attrs==17.4.0
beautifulsoup4==4.6.0
bitcointools==1.1.44
bumpversion==0.5.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
crypto==1.4.1
cytoolz==0.9.0.1
DBUtils==1.2
decorator==4.2.1
docopt==0.6.2
elasticsearch==6.2.0
elasticsearch-dsl==6.1.0
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.8.1
execnet==1.5.0
flaky==3.4.0
Flask==0.12.2
hypothesis==3.48.1
idna==2.6
ipaddress==1.0.19
ipython==5.7.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
lru-dict==1.1.6
lxml==4.2.1
MarkupSafe==1.0
Naked==0.1.31
parso==0.1.1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
py-solc==2.1.0
pycrypto==2.6.1
pycurl==7.43.0.1
Pygments==2.2.0
pymongo==3.6.1
PyMySQL==0.8.0
pysha3==1.0.2
pytest==3.2.5
pytest-forked==0.2
pytest-mock==1.7.1
pytest-pythonpath==0.7.2
pytest-watch==4.1.0
pytest-xdist==1.22.2
python-dateutil==2.7.2
python-memcached==1.59
pytz==2018.3
PyYAML==3.12
qrcode==6.0
redis==2.10.6
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
shellescape==3.4.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==0.9.3
toolz==0.9.0
tornado==4.3
tornado-redis==2.4.18
tox==2.9.1
tqdm==4.19.6
traitlets==4.3.2
tzlocal==1.5.1
ujson==1.35
urllib3==1.22
virtualenv==15.1.0
watchdog==0.8.3
wcwidth==0.1.7
web3==3.16.5
websockets==3.4
Werkzeug==0.14.1
when-changed==0.3.0
xlwt==1.3.0
pika==0.11.2